+++ backend/.gitignore
@@ -1,2 +1,3 @@
 # Test files
 src/secret_settings.py
+src/unit_tests
